Refactor code for consistency and readability
- Updated function parameter formatting for consistency across multiple files.
- Added null checks after memory allocations in Huffman and Binary Search Tree implementations.
- Improved memory management by freeing allocated memory in Making Change and Huffman test files.
- Corrected variable naming inconsistencies (RowSIze to RowSize) in MazeSolver.
- Enhanced readability by adjusting spacing and indentation in various functions.
- Removed unnecessary comments and improved existing comments for clarity.

+++ 14/Huffman/Test_Huffman.c
@@ -1,36 +1,36 @@
 #include "Huffman.h"
 #include <string.h>
 
-int main( void )
-{
-    char* Source  = "This Is Algorithms.";
+int main(void) {
+    char* Source = "This Is Algorithms.";
     char* Decoded = "";
 
     HuffmanNode* Tree = NULL;
-    BitBuffer    Encoded = {NULL, 0};
-    HuffmanCode  CodeTable[MAX_CHAR];
-    
-    memset ( &CodeTable, 0, sizeof(HuffmanCode) * MAX_CHAR );
-    
-    Huffman_Encode( &Tree, (UCHAR*)Source, &Encoded, CodeTable );
-    
-    printf("Original Size:%I64d Encoded Size:%d\n", 
-        (strlen(Source) + 1) * sizeof(char) * 8, Encoded.Size);
+    BitBuffer Encoded = {NULL, 0};
+    HuffmanCode CodeTable[MAX_CHAR];
+
+    memset(&CodeTable, 0, sizeof(HuffmanCode) * MAX_CHAR);
+
+    Huffman_Encode(&Tree, (UCHAR*)Source, &Encoded, CodeTable);
+
+    printf("Original Size:%lu Encoded Size:%u\n",
+           (unsigned long)((strlen(Source) + 1) * sizeof(char) * 8),
+           Encoded.Size);
 
     Decoded = (char*)malloc(sizeof(char) * (strlen(Source) + 1));
-    Huffman_Decode( Tree, &Encoded, (UCHAR*)Decoded );
+    Huffman_Decode(Tree, &Encoded, (UCHAR*)Decoded);
 
-    printf("Original : %s\n", Source );
+    printf("Original : %s\n", Source);
     printf("Encoded  : ");
-    
-    Huffman_PrintBinary( &Encoded );
+
+    Huffman_PrintBinary(&Encoded);
 
     printf("\n");
 
-    printf("Decoded  : %s\n", Decoded );
+    printf("Decoded  : %s\n", Decoded);
 
-    Huffman_DestroyTree( Tree );
-    free( Decoded );
+    Huffman_DestroyTree(Tree);
+    free(Decoded);
 
     return 0;
 }
